CORE-32995 Split out log and configure commands into separate modules.
